feat: new bench_rspack_commit workflows
