+++ .github/workflows/bench_rspack_commit.yml
@@ -29,42 +29,52 @@ jobs:
         shardIndex: [1, 2, 3]
         shardTotal: [3]
       fail-fast: false
-    outputs:
-      diff-result: ${{ steps.print-results.outputs.diff-result }}
     steps:
-      - name: Checkout
-        uses: actions/checkout@v2
+      - uses: actions/checkout@v4
       - name: Init env
         uses: ./.github/actions/env
       - uses: ./.github/actions/build-rspack
         with:
           path: ${{ env.RSPACK_DIR }}
       - name: Run benchmark
-        run: node bin/cli.js bench --shard ${{ matrix.shardIndex }}/${{ matrix.shardTotal }}
+        run: node bin/cli.js bench --shard ${{ matrix.shardIndex }}/${{ strategy.job-total }}
+      - uses: actions/upload-artifact@v4
+        with:
+          name: benchmark-artifacts-${{ matrix.shardIndex }}-${{ strategy.job-total }}
+          path: output
+
+  comment-compare-results:
+    runs-on: ubuntu-latest
+    needs: bench
+    if: always()
+    steps:
+      - uses: actions/checkout@v4
+      - name: Init env
+        uses: ./.github/actions/env
+      - uses: actions/download-artifact@v4
+        with:
+          pattern: benchmark-artifacts-*
+          path: output
+          merge-multiple: true
+      - shell: bash
+        run: ls -lah output
       - id: print-results
         name: Print results
         run: |
-          result=$(node bin/compare-bench.js latest current)
+          result=$(node bin/cli.js compare --base latest --current current)
           echo "$result"
           echo "diff-result=${result//$'\n'/'@@'}" >> $GITHUB_OUTPUT
           if [[ $result =~ "Threshold exceeded" ]]; then
             echo "Some benchmark cases exceed the threshold, please visit the previous step for more information"
             exit 1
           fi
-
-  create-comment:
-    runs-on: ubuntu-latest
-    needs: [run-bench]
-    if: always()
-    steps:
-      - id: create-comment
-        uses: actions/github-script@v6
+      - uses: actions/github-script@v6
         with:
           github-token: ${{ secrets.RSPACK_BOT_ACCESS_TOKEN }}
           result-encoding: string
           script: |
-            const diffResult = `${{ needs.run-bench.outputs.diff-result }}`
-            let result = "task ${{ needs.run-bench.result }}"
+            const diffResult = `${{ steps.print-results.outputs.diff-result }}`
+            let result = "task ${{ needs.bench.result }}"
             if (diffResult) {
               result = diffResult.replace(/@@/g, "\n");
             }
