+++ bin/cli.js
@@ -6,6 +6,7 @@ import { $, cd } from 'zx';
 import actionsCore from "@actions/core";
 import { run, formatResultTable } from "../lib/index.js";
 import { isGitHubActions, dirExist } from "../lib/utils.js";
+import { compare } from "../lib/compare.js";
 
 $.verbose = true
 
@@ -37,7 +38,16 @@ const cli = meow({
         shard: {
             type: 'string',
             default: '1/1'
-        }
+        },
+
+        base: {
+            type: 'string',
+            default: 'latest'
+        },
+        current: {
+            type: 'string',
+            default: 'current'
+        },
 	}
 });
 
@@ -50,7 +60,10 @@ const {
     binding,
     js,
 
-    shard
+    shard,
+
+    base,
+    current
 } = cli.flags;
 
 const cwd = process.cwd();
@@ -59,15 +72,16 @@ const configPath = join(process.cwd(), 'bench.config.js');
 const config = (await import(configPath)).default;
 
 const jobs = config.jobs ?? [];
-const rspackWorkspace = config.rspackWorkspace ?? join(cwd, '.rspack');
+const rspackDirectory = config.rspackDirectory ?? join(cwd, '.rspack');
+const benchmarkDirectory = config.benchmarkDirectory ?? join(cwd, 'output');
 
 if (!command || command === 'build') {
     const fetchUrl = `https://github.com/${repository}`;
-    if (!(await dirExist(rspackWorkspace))) {
-		await $`git clone ${fetchUrl} ${rspackWorkspace}`;
+    if (!(await dirExist(rspackDirectory))) {
+		await $`git clone ${fetchUrl} ${rspackDirectory}`;
 	}
 
-    cd(rspackWorkspace);
+    cd(rspackDirectory);
 
     await $`git reset --hard`;
     const currentBranch = (await $`git rev-parse --abbrev-ref HEAD`).toString().trim();
@@ -131,4 +145,8 @@ if (!command || command === 'bench') {
     } else {
         console.log(`No jobs to run for shard ${currentIndex}/${totalShards}.`);
     }
+}
+
+if (!command || command === 'compare') {
+    compare(base, current, benchmarkDirectory);
 }
